+++ windivert-sys/src/bindings/mod.rs
@@ -33,13 +33,16 @@ pub const WINDIVERT_PARAM_QUEUE_SIZE_MIN: u64 = 65535; /* 64KB */
 /// Maximum valid value for queue size parameter.
 pub const WINDIVERT_PARAM_QUEUE_SIZE_MAX: u64 = 33554432; /* 32MB */
 /// Maximum valid value for priority parameter.
-pub const WINDIVERT_PRIORITY_MAX: u32 = 30000;
+pub const WINDIVERT_PRIORITY_MAX: i16 = 30000;
 /// Minimum valid value for priority parameter.
-pub const WINDIVERT_PRIORITY_MIN: i32 = -30000;
+pub const WINDIVERT_PRIORITY_MIN: i16 = -30000;
 /// Maximum valid batch length.
 pub const WINDIVERT_BATCH_MAX: u32 = 255;
 /// Maximum valid mtu size.
 pub const WINDIVERT_MTU_MAX: u32 = 65575;
+
+/// Alias type for meaningful retur types
+pub type BOOL = core::ffi::c_int;
 
 extern "C" {
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_open)
@@ -57,7 +60,7 @@ extern "C" {
         packetLen: u32,
         pRecvLen: *mut u32,
         pAddr: *mut address::WINDIVERT_ADDRESS,
-    ) -> i32;
+    ) -> BOOL;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_recv_ex)
     pub fn WinDivertRecvEx(
@@ -69,7 +72,7 @@ extern "C" {
         pAddr: *mut address::WINDIVERT_ADDRESS,
         pAddrLen: *mut u32,
         lpOverlapped: *mut core::ffi::c_void,
-    ) -> i32;
+    ) -> BOOL;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_send)
     pub fn WinDivertSend(
@@ -78,7 +81,7 @@ extern "C" {
         packetLen: u32,
         pSendLen: *mut u32,
         pAddr: *const address::WINDIVERT_ADDRESS,
-    ) -> i32;
+    ) -> BOOL;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_send_ex)
     pub fn WinDivertSendEx(
@@ -90,27 +93,27 @@ extern "C" {
         pAddr: *const address::WINDIVERT_ADDRESS,
         addrLen: u32,
         lpOverlapped: *mut core::ffi::c_void,
-    ) -> i32;
+    ) -> BOOL;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_shutdown)
-    pub fn WinDivertShutdown(handle: *mut core::ffi::c_void, how: WinDivertShutdownMode) -> i32;
+    pub fn WinDivertShutdown(handle: *mut core::ffi::c_void, how: WinDivertShutdownMode) -> BOOL;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_close)
-    pub fn WinDivertClose(handle: *mut core::ffi::c_void) -> i32;
+    pub fn WinDivertClose(handle: *mut core::ffi::c_void) -> BOOL;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_set_param)
     pub fn WinDivertSetParam(
         handle: *mut core::ffi::c_void,
         param: WinDivertParam,
         value: u64,
-    ) -> i32;
+    ) -> BOOL;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_get_param)
     pub fn WinDivertGetParam(
         handle: *mut core::ffi::c_void,
         param: WinDivertParam,
         pValue: *mut u64,
-    ) -> i32;
+    ) -> BOOL;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_parse_packet)
     pub fn WinDivertHelperParsePacket(
@@ -123,11 +126,11 @@ extern "C" {
         ppIcmpv6Hdr: *mut header::PWINDIVERT_ICMPV6HDR,
         ppTcpHdr: *mut header::PWINDIVERT_TCPHDR,
         ppUdpHdr: *mut header::PWINDIVERT_UDPHDR,
-        ppData: *mut core::ffi::c_void,
+        ppData: *mut *mut core::ffi::c_void,
         pDataLen: *mut u32,
-        ppNext: *mut core::ffi::c_void,
+        ppNext: *mut *mut core::ffi::c_void,
         pNextLen: *mut u32,
-    ) -> i32;
+    ) -> BOOL;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_hash_packet)
     pub fn WinDivertHelperHashPacket(
@@ -140,27 +143,27 @@ extern "C" {
     pub fn WinDivertHelperParseIPv4Address(
         addrStr: *const core::ffi::c_char,
         pAddr: *mut u32,
-    ) -> i32;
+    ) -> BOOL;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_parse_ipv6_address)
     pub fn WinDivertHelperParseIPv6Address(
         addrStr: *const core::ffi::c_char,
         pAddr: *mut u32,
-    ) -> i32;
+    ) -> BOOL;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_format_ipv4_address)
     pub fn WinDivertHelperFormatIPv4Address(
         addr: u32,
         buffer: *mut core::ffi::c_char,
         bufLen: u32,
-    ) -> i32;
+    ) -> BOOL;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_format_ipv6_address)
     pub fn WinDivertHelperFormatIPv6Address(
         pAddr: *const u32,
         buffer: *mut core::ffi::c_char,
         bufLen: u32,
-    ) -> i32;
+    ) -> BOOL;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_calc_checksums)
     pub fn WinDivertHelperCalcChecksums(
@@ -168,10 +171,10 @@ extern "C" {
         packetLen: u32,
         pAddr: *mut address::WINDIVERT_ADDRESS,
         flags: ChecksumFlags,
-    ) -> i32;
+    ) -> BOOL;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_dec_ttl)
-    pub fn WinDivertHelperDecrementTTL(pPacket: *mut core::ffi::c_void, packetLen: u32) -> i32;
+    pub fn WinDivertHelperDecrementTTL(pPacket: *mut core::ffi::c_void, packetLen: u32) -> BOOL;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_compile_filter)
     pub fn WinDivertHelperCompileFilter(
@@ -181,7 +184,7 @@ extern "C" {
         objLen: u32,
         errorStr: *mut *const core::ffi::c_char,
         errorPos: *mut u32,
-    ) -> i32;
+    ) -> BOOL;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_eval_filter)
     pub fn WinDivertHelperEvalFilter(
@@ -189,7 +192,7 @@ extern "C" {
         pPacket: *const core::ffi::c_void,
         packetLen: u32,
         pAddr: *const address::WINDIVERT_ADDRESS,
-    ) -> i32;
+    ) -> BOOL;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_format_filter)
     pub fn WinDivertHelperFormatFilter(
@@ -197,7 +200,7 @@ extern "C" {
         layer: WinDivertLayer,
         buffer: *mut core::ffi::c_char,
         bufLen: u32,
-    ) -> i32;
+    ) -> BOOL;
 
     /// Check the official [docs](https://reqrypt.org/windivert-doc.html#divert_helper_ntoh)
     pub fn WinDivertHelperNtohs(x: u16) -> u16;